bugfix #46 - file-structure rule incompliance fix
