+++ src/app/component/Footer/Footer.component.js
