+++ src/app/component/GoogleLogIn/GoogleLogIn.config.js
