+++ src/app/component/Starters/Starters.component.js
@@ -16,7 +16,7 @@ import {
 } from './Starters.config';
 
 import './Starters.style.scss';
-import Starter from '@component/Starter/Starter';
+import Starter from '@component/Starter/Starter.component';
 
 class Starters extends React.Component {
     static propTypes = {
