+++ src/app/store/index.js
@@ -4,6 +4,25 @@
 *
 * @license MIT
 */
-// eslint-disable-next-line import/prefer-default-export
-export { buyCake } from './CakeStore/CakeStore.action';
-export { subscribeAccount, unsubscribeAccount } from './Account/Account.action';
+
+import {
+    combineReducers,
+    createStore
+} from 'redux';
+
+import menuListReducer from './MenuList/MenuList.reducer';
+import cakeReducer from './CakeStore/CakeStore.reducer';
+import accountReducer from './Account/Account.reducer';
+
+const rootReducer = combineReducers({
+    menuList: menuListReducer,
+    cakeList: cakeReducer,
+    account: accountReducer
+});
+
+const store = createStore(
+    rootReducer,
+    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+);
+
+export default store;
