--- src/app/store/store.js
@@ -1,15 +1,0 @@
-/**
-* SIA Trinat restaurant project
-* Copyright © winniepukki. All rights reserved.
-*
-* @license MIT
-*/
-import { createStore } from 'redux';
-import rootReducer from './rootReducer';
-
-const store = createStore(
-    rootReducer,
-    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-);
-
-export default store;
